Enhance JSON schema to use required and default properties
