+++ packages/config/index.ts
@@ -10,6 +10,7 @@
 */
 
 export * from "./src/Config";
+export * from "./src/ConfigSchema";
 export * from "./src/doc/IConfig";
 export * from "./src/doc/IConfigLayer";
 export * from "./src/doc/IConfigOpts";
