+++ packages/config/src/doc/IConfig.ts
@@ -12,6 +12,7 @@
 import { IConfigProfile } from "./IConfigProfile";
 
 export interface IConfig {
+    $schema?: string;
     defaults: { [key: string]: string };
     profiles: { [key: string]: IConfigProfile };
     plugins: string[];
